refactor: useCartCheckbox hook을 useCart로 변경
